refactor: 필요없는 컴포넌트 통합 후 삭제
