--- src/shared/components/TitleCategoryCommon.tsx
@@ -1,9 +1,0 @@
-const CategoryCommonTitle = () => {
-	return (
-		<header>
-			<h1 className="head-bold-24 text-gray-04">카테고리 추가</h1>
-		</header>
-	);
-};
-
-export default CategoryCommonTitle;
